requirements.txt
# Flask Web Framework
Flask>=2.3.0

# Computer Vision dan Machine Learning
opencv-python>=4.8.0
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=10.0.0

# Database
mysql-connector-python>=8.0.0

# Utilities
python-dotenv>=1.0.0
Werkzeug>=2.3.0

# Optional: untuk development
# flask-cors>=4.0.0
